split off serial ctl in preparation of fake ctl
